+++ lisp/tree-sitter.el
@@ -101,7 +101,7 @@ END is the end of the changed text."
     (setq tree-sitter-tree
           ;; https://github.com/ubolonton/emacs-tree-sitter/issues/3
           (ts--without-restriction
-            (ts-parse tree-sitter-parser #'ts-buffer-input tree-sitter-tree)))
+            (ts-parse-chunks tree-sitter-parser #'ts-buffer-input tree-sitter-tree)))
     (run-hook-with-args 'tree-sitter-after-change-functions old-tree)))
 
 (defun tree-sitter--enable ()